requirements.txt
streamlit
beautifulsoup4
requests
langchain
langchain-google-genai
faiss-cpu
google-generativeai
numpy
pandas
python-dotenv
scikit-learn
